chore: Move test functions to helper so they can be reused.
